+++ aeron-client/src/main/c/collections/aeron_int64_to_tagged_ptr_hash_map.h
@@ -52,10 +52,11 @@ aeron_int64_to_tagged_ptr_hash_map_t;
 
 inline size_t aeron_int64_to_tagged_ptr_hash_map_hash_key(int64_t key, size_t mask)
 {
-    return (key * 31) & mask;
+    return (size_t)(((uint64_t)key * 31u) & mask);
 }
 
-inline int aeron_int64_to_tagged_ptr_hash_map_init(aeron_int64_to_tagged_ptr_hash_map_t *map, size_t initial_capacity, float load_factor)
+inline int aeron_int64_to_tagged_ptr_hash_map_init(
+    aeron_int64_to_tagged_ptr_hash_map_t *map, size_t initial_capacity, float load_factor)
 {
     size_t capacity = (size_t)aeron_find_next_power_of_two((int32_t)initial_capacity);
 
@@ -167,7 +168,7 @@ inline int aeron_int64_to_tagged_ptr_hash_map_put(
     }
 
     map->entries[index].internal_flags = AERON_INT64_TO_TAGGED_PTR_VALUE_PRESENT;
-    map->entries[index].tag = tag;
+    map->entries[index].tag = (uint32_t)tag;
     map->entries[index].value = value;
 
     if (map->size > map->resize_threshold)
@@ -210,7 +211,8 @@ inline bool aeron_int64_to_tagged_ptr_hash_map_get(
     return false;
 }
 
-inline void aeron_int64_to_tagged_ptr_hash_map_compact_chain(aeron_int64_to_tagged_ptr_hash_map_t *map, size_t delete_index)
+inline void aeron_int64_to_tagged_ptr_hash_map_compact_chain(
+    aeron_int64_to_tagged_ptr_hash_map_t *map, size_t delete_index)
 {
     size_t mask = map->capacity - 1;
     size_t index = delete_index;
@@ -281,7 +283,7 @@ typedef void (*aeron_int64_to_tagged_ptr_hash_map_for_each_func_t)(void *clientd
 typedef bool (*aeron_int64_to_tagged_ptr_hash_map_predicate_func_t)(void *clientd, int64_t key, uint32_t tag, void *value);
 
 inline void aeron_int64_to_tagged_ptr_hash_map_for_each(
-        aeron_int64_to_tagged_ptr_hash_map_t *map, aeron_int64_to_tagged_ptr_hash_map_for_each_func_t func, void *clientd)
+    aeron_int64_to_tagged_ptr_hash_map_t *map, aeron_int64_to_tagged_ptr_hash_map_for_each_func_t func, void *clientd)
 {
     for (size_t i = 0; i < map->capacity; i++)
     {
