+++ aeron-client/src/main/c/concurrent/aeron_term_appender.h
@@ -90,7 +90,7 @@ inline int32_t aeron_term_appender_handle_end_of_log_condition(
         const int32_t padding_length = term_length - term_offset;
         aeron_data_header_t *header = (aeron_data_header_t *)(term_buffer->addr + term_offset);
 
-        aeron_term_appender_header_write(term_buffer, term_offset, padding_length, term_id, session_id, stream_id);
+        aeron_term_appender_header_write(term_buffer, term_offset, (size_t)padding_length, term_id, session_id, stream_id);
         header->frame_header.type = AERON_HDR_TYPE_PAD;
         AERON_PUT_ORDERED(header->frame_header.frame_length, padding_length);
     }
@@ -109,7 +109,7 @@ inline int32_t aeron_term_appender_claim(
 {
     const size_t frame_length = length + AERON_DATA_HEADER_LENGTH;
     const int64_t aligned_frame_length = AERON_ALIGN(frame_length, AERON_LOGBUFFER_FRAME_ALIGNMENT);
-    const int64_t raw_tail = aeron_term_appender_get_and_add_raw_tail(term_tail_counter, aligned_frame_length);
+    const int64_t raw_tail = aeron_term_appender_get_and_add_raw_tail(term_tail_counter, (size_t)aligned_frame_length);
     const int64_t term_offset = raw_tail & 0xFFFFFFFF;
     const int32_t term_id = aeron_logbuffer_term_id(raw_tail);
     const int32_t term_length = (int32_t)term_buffer->length;
@@ -150,7 +150,7 @@ inline int32_t aeron_term_appender_append_unfragmented_message(
 {
     const size_t frame_length = length + AERON_DATA_HEADER_LENGTH;
     const int64_t aligned_frame_length = AERON_ALIGN(frame_length, AERON_LOGBUFFER_FRAME_ALIGNMENT);
-    const int64_t raw_tail = aeron_term_appender_get_and_add_raw_tail(term_tail_counter, aligned_frame_length);
+    const int64_t raw_tail = aeron_term_appender_get_and_add_raw_tail(term_tail_counter, (size_t)aligned_frame_length);
     const int64_t term_offset = raw_tail & 0xFFFFFFFF;
     const int32_t term_id = aeron_logbuffer_term_id(raw_tail);
     const int32_t term_length = (int32_t)term_buffer->length;
@@ -200,7 +200,7 @@ inline int32_t aeron_term_appender_append_unfragmented_messagev(
 {
     const size_t frame_length = length + AERON_DATA_HEADER_LENGTH;
     const int64_t aligned_frame_length = AERON_ALIGN(frame_length, AERON_LOGBUFFER_FRAME_ALIGNMENT);
-    const int64_t raw_tail = aeron_term_appender_get_and_add_raw_tail(term_tail_counter, aligned_frame_length);
+    const int64_t raw_tail = aeron_term_appender_get_and_add_raw_tail(term_tail_counter, (size_t)aligned_frame_length);
     const int64_t term_offset = raw_tail & 0xFFFFFFFF;
     const int32_t term_id = aeron_logbuffer_term_id(raw_tail);
     const int32_t term_length = (int32_t)term_buffer->length;
@@ -382,7 +382,7 @@ inline int32_t aeron_term_appender_append_fragmented_messagev(
                 int32_t current_buffer_remaining = (int32_t)iov[i].iov_len - current_buffer_offset;
                 int32_t num_bytes = (bytes_to_write - bytes_written) < current_buffer_remaining ?
                     (bytes_to_write - bytes_written) : current_buffer_remaining;
-                memcpy(term_buffer->addr + payload_offset, iov[i].iov_base + current_buffer_offset, num_bytes);
+                memcpy(term_buffer->addr + payload_offset, iov[i].iov_base + current_buffer_offset, (size_t)num_bytes);
 
                 bytes_written += num_bytes;
                 payload_offset += num_bytes;
